use context-dependent huffman encoding
simply use the previous byte, which is the left block
on the map, as a context. depending on the context,
use a separate huffman table. as our stage data consists
of 10 kinds of pieces including blank and "END", it
means having 10 huffman tables. given that a huffman
table is about 33 bytes, it translates to about 300
bytes overhead, ignoring the increase of the decoder
logic, which i assmume is negligible.

although i feel this implementation is a bit hackish,
it actually reduces the total cart size considerably.
(52999 -> 51020 bytes)
more importantly, this reduces the size of the C
constants and thus the linear memory consumption.