score1.c: stop panning
because it doesn't seem working well on my phone. (android, firefox)
that is, the stereo speakers are placed at top-bottom rather than
left-right when the phone in the portrait mode. unfortunately, it
seems like the typical configuration for at least android phones.
for that reason, typical applications targetting to phones seem to
make the sound panned to the center. i guess stereo sound is not
a thing for phones.